pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "isthisipbad"
version = "2.0.0"
description = "Check an IP address against popular IP and DNS blacklists"
readme = "readme.md"
license = "MIT"
requires-python = ">=3.8"
authors = [
    { name = "Jerry Gamblin", email = "jgamblin@gmail.com" }
]
keywords = [
    "security",
    "ip",
    "blacklist",
    "threat-intelligence",
    "dnsbl",
    "spam",
    "reputation"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: System :: Networking",
    "Topic :: Utilities",
]
dependencies = [
    "requests>=2.28.0",
    "dnspython>=2.3.0",
    "httpx>=0.24.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
    "types-requests>=2.28.0",
]

[project.urls]
Homepage = "https://github.com/jgamblin/isthisipbad"
Repository = "https://github.com/jgamblin/isthisipbad"
Issues = "https://github.com/jgamblin/isthisipbad/issues"

[project.scripts]
isthisipbad = "isthisipbad.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["isthisipbad"]

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=isthisipbad"
